refactor: split member history view from custom-transaction handlers
